+ fix lost history
